requirements.txt
langchain==0.2.1
langchain_google_genai==1.0.5
matplotlib==3.7.2
matplotlib==3.8.0
matplotlib==3.8.3
pandas==1.5.3
pandasai==2.1
Pillow==10.0.1
Pillow==10.3.0
protobuf==5.27.0
PyPDF2==3.0.1
python-dotenv==1.0.1
Requests==2.32.3
streamlit==1.31.1
streamlit_lottie==0.0.5
streamlit_option_menu==0.3.12
~treamlit==1.30.0
